requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.20.12
botocore==1.23.12
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.0
decorator==5.1.0
Deprecated==1.2.13
Django==3.2.9
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-rest-knox==4.1.0
djangorestframework==3.12.4
dynaconf==3.1.7
hvac==0.11.2
idna==3.3
Jinja2==3.0.3
jmespath==0.10.0
knox==0.1.14
kombu==5.2.2
loguru==0.5.3
MarkupSafe==2.0.1
prompt-toolkit==3.0.23
psycopg2==2.9.2
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
requests==2.26.0
s3transfer==0.5.0
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
validators==0.18.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
